share createCompletedRequest to specHelper
