+++ packages/rum-core/src/domain/rumEventsCollection/resource/resourceCollection.spec.ts
@@ -2,11 +2,10 @@ import type { Duration, RelativeTime, ServerDuration, TimeStamp } from '@datadog
 import { isIE, RequestType, ResourceType } from '@datadog/browser-core'
 import { createResourceEntry } from '../../../../test/fixtures'
 import type { TestSetupBuilder } from '../../../../test/specHelper'
-import { setup, stubPerformanceObserver } from '../../../../test/specHelper'
+import { setup, stubPerformanceObserver, createCompletedRequest } from '../../../../test/specHelper'
 import type { RawRumResourceEvent } from '../../../rawRumEvent.types'
 import { RumEventType } from '../../../rawRumEvent.types'
 import { LifeCycleEventType } from '../../lifeCycle'
-import type { RequestCompleteEvent } from '../../requestCollection'
 import { TraceIdentifier } from '../../tracing/tracer'
 import { validateAndBuildRumConfiguration } from '../../configuration'
 import { createRumSessionManagerMock } from '../../../../test/mockRumSessionManager'
@@ -341,16 +340,3 @@ describe('resourceCollection', () => {
     })
   })
 })
-
-function createCompletedRequest(details?: Partial<RequestCompleteEvent>): RequestCompleteEvent {
-  const request: Partial<RequestCompleteEvent> = {
-    duration: 100 as Duration,
-    method: 'GET',
-    startClocks: { relative: 1234 as RelativeTime, timeStamp: 123456789 as TimeStamp },
-    status: 200,
-    type: RequestType.XHR,
-    url: 'https://resource.com/valid',
-    ...details,
-  }
-  return request as RequestCompleteEvent
-}
